+++ CHANGES.rst
@@ -12,6 +12,7 @@ Version 0.8.0 (UNRELEASED)
 - Adds a check in scheduler if at least one workflow job could be started in Kubernetes.
 - Changes workflow execution consumer to receive only one message at a time.
 - Changes workflow execution scheduler to delay rescheduled workflows and count number of retries.
+- Adds configuration environment variable to set workflow scheduling policy (``REANA_WORKFLOW_SCHEDULING_POLICY``).
 
 Version 0.7.5 (2021-04-28)
 --------------------------